refactor!(api): move away from AJAX fallback
BREAKING CHANGE: AJAX fallback is no longer supported
